pyproject.toml
[tool.pdm]
[tool.pdm.scripts]
dev = "python manage.py runserver"
runserver = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
migrate = "python manage.py migrate"
makemigrations = "python manage.py makemigrations"
shell = "python manage.py shell"
test = "python manage.py test"
startapp = "python manage.py startapp {args}"
loaddata = "python manage.py loaddata {args}"
dumpdata = "python manage.py dumpdata {args}"
[tool.black]
line-length = 120
[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.12.0",
    "black>=23.7.0",
]
[[tool.pdm.autoexport]]
filename = "requirements.txt"
without-hashes = "true"
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "João Felipi", email = "joaofelipi877@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-extensions>=3.2.3",
    "pydotplus>=2.0.2",
    "drf-spectacular>=0.26.4",
    "setuptools>=68.1.2",
    "python-magic>=0.4.27",
    "Pillow>=10.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
